Add stock transfer workflow

+++ app-modules/catalog/database/migrations/2026_02_08_195604_create_all_catalog_tables.php
@@ -105,7 +105,7 @@ return new class extends Migration
                 ->index()
                 ->constrained('catalog_bundles')
                 ->onDelete('cascade');
-            $table->unsignedInteger('quantity');
+            $table->bigInteger('quantity');
             $table->string('display_unit_code')->index();
             $table->foreign('display_unit_code')
                 ->references('code')
